requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.32.0
requests==2.32.3
pytest==8.3.4
google-generativeai==0.8.3
python-dotenv==1.0.1

